third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-summary-show-all-expected.txt

Tests Summary view of detailed heap snapshots. The "Show All" button must show all nodes.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll

Profiler was disabled.
